+++ src/view/src/rocprofvis_flame_track_item.h
@@ -44,7 +44,7 @@ class FlameTrackItem : public TrackItem
 
 public:
     FlameTrackItem(DataProvider&                      dp,
-                   std::shared_ptr<TimelineSelection> timeline_selection, int chart_id,
+                   std::shared_ptr<TimelineSelection> timeline_selection, uint64_t chart_id,
                    std::string name, float zoom, double movement, double min_x,
                    double max_x, double scale_x, float level_min, float level_max);
     ~FlameTrackItem();
